Implemented semantic admin with password protection for DELETE routes.

+++ backend/src/routes.ts
@@ -88,15 +88,23 @@ async function DoggrRoutes(app: FastifyInstance, _options = {}) {
 	});
 	
 	// DELETE
-	app.delete<{ Body: { email } }>("/users", async (req, reply) => {
-		const {email} = req.body;
+	app.delete<{ Body: { email, password } }>("/users", async (req, reply) => {
+		const {email, password} = req.body;
 		
 		try {
-			const theUser = await req.em.findOne(User, {email});
-			
-			await req.em.remove(theUser).flush();
-			console.log(theUser);
-			reply.send(theUser);
+		  const theUser = await req.em.findOne(User, {email},
+				{populate: ['sent_messages', 'recieved_messages', 'matches', 'matched_by']});
+		  
+		  // check for admin password
+		  if (process.env.ADMIN_PASS === password) {
+		  	await req.em.remove(theUser).flush();
+		  	console.log(theUser);
+		  	reply.send(theUser);
+		  } else {
+				const error = "Admin password is missing or incorrect. User could not be deleted.";
+				console.log(error);
+				reply.status(401).send(error);
+		  }
 		} catch (err) {
 			console.error(err);
 			reply.status(500).send(err);
@@ -185,15 +193,23 @@ async function DoggrRoutes(app: FastifyInstance, _options = {}) {
 	});
 	
 	// DELETE a specific message
-	app.delete<{ Body: { messageId } }>("/messages", async (req, reply) => {
-		const {messageId} = req.body;
+	app.delete<{ Body: { messageId, password } }>("/messages", async (req, reply) => {
+		const {messageId, password} = req.body;
 		
 		try {
 			const theMessage = await req.em.findOne(Message, {messageId});
 			
-			await req.em.remove(theMessage).flush();
-			console.log(theMessage);
-			reply.send(theMessage);
+		  // check for admin password
+		  if (process.env.ADMIN_PASS === password) {
+		  	await req.em.remove(theMessage).flush();
+		  	console.log(theMessage);
+		  	reply.send(theMessage);
+		  } else {
+				const error = "Admin password is missing or incorrect. Message could not be deleted.";
+				console.log(error);
+				reply.status(401).send(error);
+		  }
+			
 		} catch (err) {
 			console.error(err);
 			console.log("Message could not be deleted");
@@ -202,15 +218,23 @@ async function DoggrRoutes(app: FastifyInstance, _options = {}) {
 	});
 	
 	// DELETE all messages user has sent
-	app.delete<{ Body: { sender: string } }>("/messages/all", async (req, reply) => {
-		const {sender} = req.body;
+	app.delete<{ Body: { sender: string, password } }>("/messages/all", async (req, reply) => {
+		const {sender, password} = req.body;
 		
 		try {
 			const theUser = await req.em.findOne(User, {email: sender});
-			await req.em.nativeDelete(Message, {sender: theUser});
-			await req.em.flush();
-			console.log(theUser);
-			reply.send(theUser);
+			
+		  	// check for admin password
+		  	if (process.env.ADMIN_PASS === password) {
+			  await req.em.nativeDelete(Message, {sender: theUser});
+			  await req.em.flush();
+			  console.log(theUser);
+			  reply.send(theUser);
+		  	} else {
+				const error = "Admin password is missing or incorrect. Messages could not be deleted.";
+				console.log(error);
+				reply.status(401).send(error);
+		  	}
 		} catch (err) {
 			console.error(err);
 			reply.status(500).send(err);
